2dcontext/scroll/scrollPathIntoView.verticalRL.html

<!DOCTYPE html>
<html style="writing-mode: vertical-rl;">
<head>
<meta charset="UTF-8">
<title>scrollPathIntoView in vertical-rl writing mode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
</head>
<body>
<div id="log"></div>
<canvas id="c" width="300" height="150" style="position: absolute; right: 100px; top: 200px;"></canvas>
<div style="width: 200vw; height: 200vh;"></div>
<script>
var t = async_test('');
_addTest(function (canvas, ctx) {
    window.scrollTo(0, 0);
    ctx.beginPath();
    ctx.rect(10, 20, 40, 80);
    ctx.scrollPathIntoView();
    var rect = c.getBoundingClientRect();
    var viewportWidth = document.scrollingElement.clientWidth;
    assert_approx_equals(rect.right, viewportWidth + (300 - 10 - 40), 0.5, 'The canvas right edge is on 250px right of the viewport right edge');
    assert_approx_equals(rect.top, 200, 0.5, 'The canvas top edge is 200px below the viewport top edge');
});
</script>
</body>
</html>
